move track & playlist structs to state.rs
